git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@171 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocserver/internals/functions/rec_rubrics/putRecIntoRubric.sql
@@ -1,4 +1,4 @@
-create or replace function putRecIntoRubric (int4, int4) returns int4 as
+create or replace function putRecIntoRubric (int8, int4) returns int4 as
 $BODY$
 declare
     idRecord alias for $1;
